requirement.txt
Django==4.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
